arm: break and distribute literal pool
c4.c needs more than 4096 bytes for literal pool (with duplicates)
and makes it hard to access literals from the pool's base pointer.
Destroy the global pool and add the pool at the end of each
function and uses pc as the pool's base pointer.

If a function gets too large that the pool at the end of the
function is out of reach, split the pool again and embed it in
the middle of instruction stream. Just make sure we don't upset
the decoding phase in the pipeline.